Added Boss/UI
I added the boss in the scene over where the boss is in the rough map. I put the UI I made. I know we don't want a coin counter, but I added it for now, so we have something. I made a little tweaks to the player, and added Text Mesh Pro. I also set up basic player health, enemy, and coin collectibles.