Fix 5 Closed Bugs
From the worksheet, I made the camera follow the player, I created the death system so that the player can actually die and respawn, and created death falling, I have yet to upload the idle animation, but I filled out the word doc as if I did so it is good to go on that part for turn in. I finished filling out the 5 Closed Issues part of the doc.