Did a lot
I added a tilemap and created the entire layout of the game with it (following the rough map). I added/fixed the player walk/idle/jump anims. I also populated the area with health and enemy placeholders as well as coins. Now, we just need enemy sprites, a moving platform, and a tutorial in the beginning.